Remove unused option split code
This code was an attempt at modeling option splits. The reality of the situation
is far more complicated and requires an external data source to be properly modeled.
Some things to consider are handling a split event before all the contracts from the
previous split event have expired.

With respect to the holdings, I think a more generic method to rewrite the holdings
would be beneficial when we have the data, but it isn't as simple as applying a split
factor. During splits and other corporate events, the OCC will rewrite the options
conract such that you are 'made whole'. Sometimes this involves adjusting the contract
multiplier away from the default 100 (or 10 for minis). Another fairly common case
(especially with M&A's) is to have part of the contract paid out in cash. So it's
possible for pretty much every value in the contract/holdings to change:

	1. Multipler
	2. Strike Price
	3. Number of contracts held
	4. Cash payout in leiu of fractional stock
	5. Constant cash payout