Fix stream dispose in ZipDataCacheProvider
This bug was causing all zip files to remain open, even when removing them from the cache and calling Dispose on ZipFile objects.

The issue is that ZipFile.Dispose closes the underlying FileStream only if created by the ZipFile instance, which is not the case here because we fetch a Stream object containing the zip file from the data provider. In fact calling ZipFile.Read(stream) returns a ZipFile instance which does not "own" the stream, it is the caller's responsibility to dispose of it.