Check SecurityChanges when ensuring currency data feeds
A performance improvement to the way we process currency conversion data
began saving the security reference instead of just the symbol. When added
via universe selection (not AddSecurity) the security doesn't exist in the
security manager until it's added in the algorithm manager. This was caused
key not found exceptions when attempting to retrieve the security from the
security manager.

The fix here is to also check SecurityChanges when resolving the data feed.
A future improvement might be to decouple the cash object from the currency
feed required to provide it conversions to the account currency.

Fixes #1635