Prevent adding universe securities to SecurityChanges in live mode
The LiveTradingDataFeed was not checking if a subscription is for a universe or not,
this led to us adding universe securities to the SecurityChanges collection. This
fix simply checks if the subscription is for universe selection, and if it is, does
not add the security to the SecurityChanges collection.

Fixes #1656