Send finalized insights to messaging handler
The AlphaResultPacketSender is implemented as an IInsightManagerExtension.
It binds to the event that fires when an insight is done being scored.
These are enqueued for later pushing via the messaging handler. Nominally
the limits are set to emit a maximum of 50 finalized insights, with a
second between the completion of a send to the start of the next send
(timer disabled while sending).