Add Contracts function to FutureFilterUniverse
FutureFilterUniverse was lacking a means of explicitly setting the set of
contract symbols to be selected, which means algorithms had to use the
vairous helper methods but wereunable to define their own selection function
for the ontract universe. This change adds two methods, one to explicitly
set the contracts to be selected and another that mirrors the way universe
selection works, by providing a function that accept sthe full set of contract
symbols and returns the filtered set of contract symbols. This enables LINQ
filtering of the contract universe.