Change date range of daily regression
Since we start Monday and end on a Friday we aren't properly covering the
edge cases. As it currently stands, if we start on a date then we receive
data for that date, so a daily algo starting on the 8th receives it's first
data point on the 9th at 00:00 (daily bar is 8th 00:00 to 9th 00:00). Like
wise, the requested end date is the 17th and we include data from the 17th,
so the last data point is the 18th at 00:00 (daily bar is 17th 00:00 to 18th
00:00).