Add Task extension methods for safe synchronous await
Task.Result has many dangers surrounding the original captured SynchronizationContext.
Issues arise when the original thread context is also waiting on the task to complete
causing a deadlock. In 99% of cases in library code, there's no need to utilize the
SyncronizationContext for executing delegates in the calling thread's context. This
viewpoint is heavily supported by MSDN literate and various programming forums alike:

https://msdn.microsoft.com/en-us/magazine/gg598924.aspx
https://msdn.microsoft.com/en-us/magazine/jj991977.aspx
https://medium.com/bynder-tech/c-why-you-should-use-configureawait-false-in-your-library-code-d7837dce3d7f
https://stackoverflow.com/questions/27851073/why-would-i-bother-to-use-task-configureawaitcontinueoncapturedcontext-false