Prevent overwriting computed fees for backwards compatability
For years IFillModel has been responsibile for computing an order's fee. The
standard way of doing this has always been via invocation of security.FeeModel,
but there was no mechanism preventing an  implementation from directly
computing the fee. As such, this check was added to prevent overwriting an order
fee that has already been computed.