Increment Intrinio time between calls to 1 minute
The actual implementation uses the Intrinio `historical_data` end point and ask for CSV format. At the moment of development, the key they provided and the one used in the test has a limit of 1 call per second for the historical_data` end point`, now the key only free access 1 call per minute.

This commit is the fastest fix for the test failing but is impractical for any user with a free account. Maybe a better solution is to implement `historical_data` end point and but asking for the JSON (default) format. This implies implementing reading the data from JSON and implement paging in the `BaseData.Read` method.