Reset error flags before restarting IBGateway
All these flags are now being reset for consistency before restarting but the only one considered a bug fix is for `_previouslyInResetTime`.

This flag is normally set to `true` in `TryWaitForReconnect` after a disconnect but it was never being cleared after an IBGateway restart, causing an unnecessary reboot outside of the IB reset period on the next `TryWaitForReconnect` call (which in turn could also cause algorithm termination later on at cash sync time).